# codegen/swagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "TODO: Add a description"
  version: "1.0"
  title: "Tesla Model S JSON API"
host: "owner-api.teslamotors.com"
basePath: "/"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /api/1/vehicles/{vehicle_id}/wake_up:
    post:
      tags:
      - "Vehicle Commands"
      summary: "Wake Up Car"
      description: "Wakes up the car from the sleep state. Necessary to get some data\
        \ from the car."
      operationId: "WakeUpVehicle"
      produces:
      - "application/json"
      parameters:
      - name: "vehicle_id"
        in: "path"
        description: "The id of the Vehicle."
        required: true
        type: "string"
        x-exportParamName: "VehicleId"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/WakeUpResponse"
...
  /api/1/vehicles/{vehicle_id}/command/charge_port_door_open:
    post:
      tags:
      - "Vehicle Commands"
      summary: "Open Charge Port"
      description: "Opens and unlocks the charge port."
      operationId: "OpenChargePort"
      produces:
      - "application/json"
      parameters:
      - name: "vehicle_id"
        in: "path"
        description: "The id of the Vehicle."
        required: true
        type: "string"
        x-exportParamName: "VehicleId"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/CommandResponse"
  /api/1/vehicles/{vehicle_id}/command/charge_port_door_close:
    post:
      tags:
      - "Vehicle Commands"
      summary: "Close Charge Port"
      description: "Closes the charge port."
      operationId: "CloseChargePort"
      produces:
      - "application/json"
      parameters:
      - name: "vehicle_id"
        in: "path"
        description: "The id of the Vehicle."
        required: true
        type: "string"
        x-exportParamName: "VehicleId"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/CommandResponse"
...
definitions:
...
  WakeUpResponse:
    type: "object"
    properties:
      response:
        $ref: "#/definitions/WakeUpResponse_response"
    title: "Wake Up Response"
    example:
      response:
        in_service: true
        color: "color"
        calendar_enabled: true
        display_name: "display_name"
        api_version: "api_version"
        backseat_token: "backseat_token"
        backseat_token_updated_at: 1.46581298050294517310021547018550336360931396484375
        user_id: 0.80082819046101150206595775671303272247314453125
        vin: "vin"
        tokens:
        - "tokens"
        - "tokens"
        state: "state"
        id: "id"
        vehicle_id: 6.02745618307040320615897144307382404804229736328125
        option_codes: "option_codes"
  CommandResponse:
    type: "object"
    required:
    - "response"
    properties:
      response:
        $ref: "#/definitions/CommandResponse_response"
    title: "CommandResponse"
    example:
      response:
        result: true
        reason: "reason"
...
  WakeUpResponse_response:
    type: "object"
    properties:
      user_id:
        type: "number"
      vehicle_id:
        type: "number"
      vin:
        type: "string"
      display_name:
        type: "string"
      option_codes:
        type: "string"
      color:
        type: "string"
      tokens:
        type: "array"
        items:
          type: "string"
      state:
        type: "string"
      in_service:
        type: "boolean"
      id:
        type: "string"
      calendar_enabled:
        type: "boolean"
      api_version:
        type: "string"
      backseat_token:
        type: "string"
      backseat_token_updated_at:
        type: "number"
    example:
      in_service: true
      color: "color"
      calendar_enabled: true
      display_name: "display_name"
      api_version: "api_version"
      backseat_token: "backseat_token"
      backseat_token_updated_at: 1.46581298050294517310021547018550336360931396484375
      user_id: 0.80082819046101150206595775671303272247314453125
      vin: "vin"
      tokens:
      - "tokens"
      - "tokens"
      state: "state"
      id: "id"
      vehicle_id: 6.02745618307040320615897144307382404804229736328125
      option_codes: "option_codes"
  CommandResponse_response:
    type: "object"
    properties:
      result:
        type: "boolean"
      reason:
        type: "string"
    description: ""
    example:
      result: true
      reason: "reason"
...